Fix theme customization
