+++ src/types/muiMaterial.d.ts
@@ -1,0 +1,23 @@
+export * from '@mui/material';
+
+declare module '@mui/material' {
+  interface Color {
+    0?: string;
+    // 50: string;
+    // 100: string;
+    // 200: string;
+    // 300: string;
+    // 400: string;
+    // 500: string;
+    // 600: string;
+    // 700: string;
+    // 800: string;
+    // 900: string;
+    A50?: string;
+    // A100: string;
+    // A200: string;
+    // A400: string;
+    // A700: string;
+    A800?: string;
+  }
+}
