+++ src/types/muiMaterialStyles.d.ts
@@ -1,0 +1,41 @@
+export * from '@mui/material/styles';
+
+declare module '@mui/material/styles' {
+  interface PaletteColor {
+    lighter?: string;
+    100?: string;
+    200?: string;
+    // light: string;
+    400?: string;
+    // main: string;
+    600?: string;
+    // dark: string;
+    700?: string;
+    800?: string;
+    darker?: string;
+    900?: string;
+    A100?: string;
+    A200?: string;
+    A300?: string;
+    // contrastText: string
+  }
+
+  interface SimplePaletteColorOptions {
+    lighter?: string;
+    100?: string;
+    200?: string;
+    // light: string;
+    400?: string;
+    // main: string;
+    600?: string;
+    // dark: string;
+    700?: string;
+    800?: string;
+    darker?: string;
+    900?: string;
+    A100?: string;
+    A200?: string;
+    A300?: string;
+    // contrastText: string
+  }
+}
